+++ src/components/Header/StatsBox/StatsBox.tsx
@@ -3,16 +3,18 @@ import { useAllBlocks } from '@services/api/hooks/useBlocks';
 import { formatHash } from '@utils/helpers';
 import StatsDesktop from './StatsDesktop/StatsDesktop';
 import StatsMobile from './StatsMobile/StatsMobile';
-import { useMainStore } from '@services/stores/useMainStore';
 
-function StatsBox() {
+interface StatsBoxProps {
+  variant: 'desktop' | 'mobile';
+}
+
+function StatsBox({ variant }: StatsBoxProps) {
   const { data } = useAllBlocks();
   const values = data?.blockTimes || [];
   const sum = values.reduce(
     (accumulator: number, currentValue: number) => accumulator + currentValue,
     0
   );
-  const isMobile = useMainStore((state) => state.isMobile);
 
   const latestMoneroHashRate = data?.currentMoneroHashRate ?? 0;
   const latestShaHashRate = data?.currentShaHashRate ?? 0;
@@ -25,7 +27,7 @@ function StatsBox() {
   const formattedMoneroHashRate = formatHash(latestMoneroHashRate);
   const formattedSha3HashRate = formatHash(latestShaHashRate);
 
-  return isMobile ? (
+  return variant === 'mobile' ? (
     <StatsMobile
       moneroHashRate={formattedMoneroHashRate}
       shaHashRate={formattedSha3HashRate}
