--- src/components/Header/StatsMobile/StatsBox.tsx
@@ -1,35 +1,0 @@
-import numeral from 'numeral';
-import { useAllBlocks } from '@services/api/hooks/useBlocks';
-import { formatHash } from '@utils/helpers';
-import StatsMobile from './StatsMobile/StatsMobile';
-
-function StatsBox() {
-  const { data } = useAllBlocks();
-  const values = data?.blockTimes || [];
-  const sum = values.reduce(
-    (accumulator: number, currentValue: number) => accumulator + currentValue,
-    0
-  );
-
-  const latestMoneroHashRate = data?.currentMoneroHashRate ?? 0;
-  const latestShaHashRate = data?.currentShaHashRate ?? 0;
-
-  const average = sum / values.length;
-  const formattedAverageBlockTime = numeral(average).format('0') + 'm';
-  const formattedBlockHeight = numeral(
-    data?.tipInfo.metadata.best_block_height
-  ).format('0,0');
-  const formattedMoneroHashRate = formatHash(latestMoneroHashRate);
-  const formattedSha3HashRate = formatHash(latestShaHashRate);
-
-  return (
-    <StatsMobile
-      moneroHashRate={formattedMoneroHashRate}
-      shaHashRate={formattedSha3HashRate}
-      averageBlockTime={formattedAverageBlockTime}
-      blockHeight={formattedBlockHeight}
-    />
-  );
-}
-
-export default StatsBox;
