--- src/components/Header/StatsMobile/StatsMobile/StatsItem/StatsItem.tsx
@@ -1,20 +1,0 @@
-import {
-  StyledBox,
-  ValueTypography,
-  LabelTypography,
-} from './StatsItem.styles';
-
-interface Props {
-  label: React.ReactNode;
-  value: string;
-  lowerCase?: boolean;
-}
-
-export default function StatsItem({ label, value, lowerCase }: Props) {
-  return (
-    <StyledBox>
-      <ValueTypography lowerCase={lowerCase}>{value}</ValueTypography>
-      <LabelTypography>{label}</LabelTypography>
-    </StyledBox>
-  );
-}
