+++ src/theme/MainLayout.tsx
@@ -30,7 +30,8 @@ import Header from '@components/Header/Header';
 import VersionInfo from '@components/VersionInfo/VersionInfo';
 import Banner from '@components/Banner/Banner';
 import { useMainStore } from '@services/stores/useMainStore';
-import StatsBox from '@components/Header/StatsMobile/StatsBox';
+// import StatsBox from '@components/Header/StatsBox/StatsMobile/StatsBox';
+import StatsBox from '@components/Header/StatsBox/StatsBox';
 
 export default function MainLayout() {
   const isMobile = useMainStore((state) => state.isMobile);
@@ -45,8 +46,8 @@ export default function MainLayout() {
             <Outlet />
             <VersionInfo />
           </Container>
-          {isMobile && <StatsBox />}
         </Grid>
+        {isMobile && <StatsBox variant="mobile" />}
       </ThemeProvider>
     </>
   );
